requirements.txt
flask==3.0.0
flask-cors==4.0.0

python-dotenv==1.0.1

pandas==2.2.2
openpyxl==3.1.5
xlrd==2.0.1

selenium==4.18.1
webdriver-manager==4.0.1

APScheduler==3.10.4

cryptography==42.0.5
bcrypt==4.1.2
PyJWT==2.8.0

pyautogui==0.9.54
